% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catHoriz}
\alias{catHoriz}
\title{Concatenate The Coregenes}
\usage{
catHoriz(rn, ogsDirNam, afi, extension = "_supergene.fasta", n_threads = 1L)
}
\arguments{
\item{rn}{The isolates names.}

\item{ogsDirNam}{Where to put the files.}

\item{afi}{A \code{character()} vector with the paths to the orthologous
alignments.}

\item{extension}{An extension.}

\item{n_threads}{The number of cpus to use.}
}
\value{
The alignment files and the names of those files in the R stout.
}
\description{
Concatenate the core genes and produce one file per genome with
the supergene in fasta format.
}
\author{
Ignacio Ferres
}
