% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catVert}
\alias{catVert}
\title{Concatenate All Supergenes In A Single Fasta File}
\usage{
catVert(outfile, sos)
}
\arguments{
\item{outfile}{The name of the final alignment.}

\item{sos}{A \code{character()} vector with the file names of the supergenes
fasta files.}
}
\value{
A fasta file and the nameo of it in the R stdout.
}
\description{
Concatenate all supergenes in a single fasta file.
}
\author{
Ignacio Ferres
}
