% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_Handling.R
\name{comp}
\alias{comp}
\title{Complements a nucleic acid sequence}
\usage{
comp(seq, forceToLower = TRUE, ambiguous = FALSE)
}
\arguments{
\item{seq}{See \link[seqinr]{comp}.}

\item{forceToLower}{See \link[seqinr]{comp}.}

\item{ambiguous}{See \link[seqinr]{comp}.}
}
\value{
See \link[seqinr]{comp}.
}
\description{
Sligthly modified version of \link[seqinr]{comp}, from
\code{seqinr} package.
}
