% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePhylo.R
\name{computePhylo}
\alias{computePhylo}
\title{Compute Phylogeny}
\usage{
computePhylo(ali, mode = "ml", nbs = 100L, n_threads = 1,
  outPrefix = "phylo", outDir = ".", rearrangement = "NNI", ...)
}
\arguments{
\item{ali}{The path to the alignment file.}

\item{mode}{One of "nj" (Neighbour-joining) of "ml" (Maximum likelihood).}

\item{nbs}{The number of bootstraps.}

\item{n_threads}{The number of cpus to use. Just used if nbs > 0.}

\item{outPrefix}{A prefix to use on the output files.}

\item{outDir}{Where to put the newick files.}

\item{rearrangement}{Choice of tree rearrangement, only used if mode="ml".}

\item{...}{Further arguments to pass to \link[phangorn]{optim.pml}.}
}
\value{
An object of class "phylo" (phangorn and ape packages), and the
specified newick files.
}
\description{
Takes the core genes alignment and produce a phylogeny.
}
\details{
Takes the core gene alignment and produces newick files. If \code{mode}
is set to "nj", the outfile will have a "_nj.nwk" subffix. If set to "ml" and
nbs = 0L, then it will be "_ml.nwk", and if "ml" and nbs > 0L, then it will
output a second file with the bootstrap trees with the suffix
"_ml_\code{nbs}.nwk".

Multiple cpus can be used when performing bootstrap.

It uses the "GTR" model to optimize the tree.
}
\references{
Paradis E., Claude J. & Strimmer K. 2004. APE: analyses of phylogenetics and
evolution in R language. Bioinformatics 20: 289-290.
Schliep K.P. 2011. phangorn: phylogenetic analysis in R. Bioinformatics, 27(4)
592-593.
}
\author{
Ignacio Ferres
}
