% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAlign.R
\name{coreAlign}
\alias{coreAlign}
\title{Compute Core Genome Alignment}
\usage{
coreAlign(gffs = character(), hmm = character(),
  outfile = "coregenome.aln", level = 1, n_threads = 1L)
}
\arguments{
\item{gffs}{A \code{character} vector with the gff file paths.}

\item{hmm}{A \code{character} string with the path to the \code{.hmm.tar.gz}
file downloaded from EggNOG database website.}

\item{outfile}{A \code{character} string with the coregenome alignment file
name. (Default: coregenome.aln).}

\item{level}{\code{numeric}}

\item{n_threads}{\code{integer} The number of cpus to use.}
}
\value{
A core genome alignment file.
}
\description{
Identify and align core genes, and concatenate the alignments
in a single file suitable for phylogenetic analyses.
}
\author{
Ignacio Ferres
}
