% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog_interface.R
\name{download_nog_hmm}
\alias{download_nog_hmm}
\title{Download specific hmm datasets from EggNOG.}
\usage{
download_nog_hmm(nog.prefix = "proNOG", onDir = ".")
}
\arguments{
\item{nog.prefix}{\code{character}, a nog prefix as returned by
\link{list_eggnogdb} function.}

\item{onDir}{The directory name where to download the file.}
}
\value{
A tar.gz compressed file on the specified directory. On console it
returns the path to the downloaded file.
}
\description{
Takes a nog prefix and downloads a compressed (.tar.gz) file
from EggNOG db containing the hmm models of the specified group.
}
\author{
Ignacio Ferres
}
