% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_Handling.R
\name{extractGffTable}
\alias{extractGffTable}
\title{Extract the gff3 table and make it 'R'eadable.}
\usage{
extractGffTable(rl)
}
\arguments{
\item{rl}{\code{character}. A vector of character strings as passed by
\code{readLines()}, reading the gff3 file.}
}
\value{
A \code{data.frame}.
}
\description{
Read a gff3 file and transforms the table in a \code{data.frame}.
}
\author{
Ignacio Ferres
}
