% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_Handling.R
\name{extractSeqsFromGff3}
\alias{extractSeqsFromGff3}
\title{Extract Sequences from Gff3 file}
\usage{
extractSeqsFromGff3(infile, in.path = NULL, keep = "aa",
  write.in.path = "dna")
}
\arguments{
\item{infile}{\code{character}. The gff3 filename.}

\item{in.path}{Where the output (fasta files) will be written.}

\item{keep}{\code{character}. What to keep in memory. One of "aa", "dna",
"both" or "none".}

\item{write.in.path}{\code{character}. What to write \code{in.path}. One of
"aa", "dna", "both" or "none".}
}
\value{
A \code{list} with CDS gene and/or protein sequences, if specifyied.
}
\description{
Extract Sequences from Gff3 file and outputs a list of element,
one per each CDS. Each element contain the DNA gene sequence and the AA
protein sequence.
}
\author{
Gregorio Iraola and Ignacio Ferres.
}
