% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIdsFromStats}
\alias{getIdsFromStats}
\title{Get Pfam Ids From Hmmstats output}
\usage{
getIdsFromStats(stats)
}
\arguments{
\item{stats}{The path to the file where the stats were written.}
}
\value{
A \code{character} vector with the Pfam ids.
}
\description{
Get Pfam-A Ids from \code{hmmstat} output.
}
