% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{hmmPress}
\alias{hmmPress}
\title{hmmPress}
\usage{
hmmPress(model)
}
\arguments{
\item{model}{\code{character} The name of the hmm file.}
}
\value{
The names of indexed files.
}
\description{
Wrapper function of \code{hmmpress} (HMMER 3).
}
