% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{hmmStat}
\alias{hmmStat}
\title{Compute Hmm Stats}
\usage{
hmmStat(hmmfile)
}
\arguments{
\item{hmmfile}{\code{character} The path to the hmm file.}
}
\value{
The file name of the temporary file where stats were written.
}
\description{
Compute hmm stats.
}
