% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggnog_interface.R
\name{list_eggnogdb}
\alias{list_eggnogdb}
\title{List available EggNOG db datasets}
\usage{
list_eggnogdb()
}
\value{
A \code{data.frame} with the available datasets and their respective
nog prefixes.
}
\description{
This function list all the available EggNOG datasets to
download, associated with their nog prefix.
}
\author{
Ignacio Ferres
}
