% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mafft}
\alias{mafft}
\title{Align with MAFFT aligner}
\usage{
mafft(infile, outfile, n_threads = 1L, type = "nuc", mode = "linsi")
}
\arguments{
\item{infile}{\code{character()} The path to the infile.}

\item{outfile}{\code{character()} The outfile name. If missing, the
extension is changed to '.ali'.}

\item{n_threads}{\code{integer()} The number of cpus to use.}

\item{type}{One of "nuc" or "amino".}

\item{mode}{Accuracy. One of "mafft", "ginsi", "linsi" or "einsi". The first
one is the default MAFFT mode, very fast. The second uses mafft options
"--maxiterate 1000 --globalpair". The third uses "--maxiterate 1000
--localpair" (DEFAULT). The fourth uses "--ep 0 --maxiterate 1000
--genafpair". See MAFFT manual for more details.}
}
\value{
The alignment file and the name of the outfile in the R console.
}
\description{
A wrapper function to run MAFFT.
}
