% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{readDomtblout}
\alias{readDomtblout}
\title{Process hmmsearch output to make it 'R'eadable.}
\usage{
readDomtblout(domtblout)
}
\arguments{
\item{domtblout}{\code{character}. hmmsearch output file, formated as a
\code{--domtblout} table. See HMMER 3.1b2.}
}
\value{
A \code{data.frame} with the hmmsearch output.
}
\description{
Process hmmsearch output to make it readable by R.
}
\note{
Taken and adapted from \code{micropan} package (Lars Snipen and
Kristian Hovde Liland).
}
