% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setHMM}
\alias{setHMM}
\alias{setHmm}
\title{Prepare Compressed Hmm Files To Hmmsearch}
\usage{
setHmm(hmmTarGz)
}
\arguments{
\item{hmmTarGz}{The path to the \code{.hmm.tar.gz} file downloaded from
EggNOG website.}
}
\value{
A \code{character} vector with the names of the hmm file and their
indices.
}
\description{
Decompress, concatenate and press (\code{hmmpress}) hmm files
to allow \code{hmmsearch} to run.
}
\author{
Ignacio Ferres
}
