% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3_Handling.R
\name{translate}
\alias{translate}
\title{Translate}
\usage{
translate(seq, frame = 0, sens = "F", numcode = 1, NAstring = "X",
  ambiguous = FALSE)
}
\arguments{
\item{seq}{See \link[seqinr]{translate}.}

\item{frame}{See \link[seqinr]{translate}.}

\item{sens}{See \link[seqinr]{translate}.}

\item{numcode}{See \link[seqinr]{translate}.}

\item{NAstring}{See \link[seqinr]{translate}.}

\item{ambiguous}{See \link[seqinr]{translate}.}
}
\value{
See \link[seqinr]{translate}.
}
\description{
Slightly modified version of \link[seqinr]{translate} function
of the \code{seqinr} package.
}
