% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{untargz}
\alias{untargz}
\title{Decompress A Tar.gz File}
\usage{
untargz(targzfile, exdir)
}
\arguments{
\item{targzfile}{The full path to the file to decompress.}

\item{exdir}{The directory to put the decompressed files.}
}
\value{
A \code{character} vector with the decompressed files.
}
\description{
Decompress a \code{.tar.gz} file.
}
\author{
Ignacio Ferres
}
