% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAlign.R
\name{coreAlign}
\alias{coreAlign}
\title{Compute Core Genome Alignment}
\usage{
coreAlign(gffs = character(), hmmFile = character(), isCompressed = TRUE,
  outfile = "coregenome.aln", mafftMode = "linsi", ogsDirNam,
  keepOgs = FALSE, level, n_threads = 1L)
}
\arguments{
\item{gffs}{A \code{character} vector with the gff file paths.}

\item{hmmFile}{The path to the \code{.hmm.tar.gz} file downloaded from
EggNOG website, or the already prepared \code{hmm} text file. See
\code{isCompressed} below.}

\item{isCompressed}{\code{logical()} If the \code{hmm} param points to the
"hmm.tar.gz" file, it should be set to \code{TRUE}. If the pipeline has been
already ran, the \code{hmm} parameter should point to the ".hmm" file that
was generated before, and the \code{isCompressed} param should be set to
\code{FALSE}. If the pipeline was ran before, the function will also check
for index files and will produce them if any of the required is missing. See
"hmmpress" from HMMER 3.1b2 manual.}

\item{outfile}{A \code{character} string with the coregenome alignment file
name. (Default: coregenome.aln).}

\item{mafftMode}{Alignment accuracy. One of "mafft", "ginsi", "linsi" or
"einsi". The first one is the default MAFFT mode, very fast. The second uses
mafft options "--maxiterate 1000 --globalpair". The third uses "--maxiterate
1000 --localpair" (phylen DEFAULT). The fourth uses "--ep 0 --maxiterate
1000 --genafpair". See MAFFT manual for more details.}

\item{ogsDirNam}{\code{character()} The directory name where to put
intermediate files. If not provided or already exists, it will be
automatically generated. This directory is removed at the end of the
pipeline, unless \code{keepOgs = TRUE} (see below).}

\item{keepOgs}{\code{logical()} If want to keep an intermediate directory
fasta files containing the orthologous groups (DEFAULT: \code{FALSE}).}

\item{level}{\code{numeric}}

\item{n_threads}{\code{integer} The number of cpus to use.}
}
\value{
A core genome alignment file.
}
\description{
Identify and align core genes, and concatenate the alignments
in a single file suitable for phylogenetic analyses.
}
\author{
Ignacio Ferres
}
